% Auto-generated: do not edit by hand
\name{simpleGameRow}

\alias{simpleGameRow}

\title{SimpleGameRow component}

\description{

}

\usage{
simpleGameRow(index=NULL, isEditMode=NULL, onCancel=NULL, onDelete=NULL,
onEdit=NULL, onSave=NULL, p1Color=NULL, p2Color=NULL,
rowData=NULL, tagOptions=NULL)
}

\arguments{
\item{index}{Numeric. }

\item{isEditMode}{Logical. }

\item{onCancel}{}

\item{onDelete}{}

\item{onEdit}{}

\item{onSave}{}

\item{p1Color}{Character. }

\item{p2Color}{Character. }

\item{rowData}{Lists containing elements 'winner', 'first', 'p1tags', 'p2tags'.
those elements have the following types:
  - winner (character; optional)
  - first (character; optional)
  - p1tags (list of characters; optional)
  - p2tags (list of characters; optional). }

\item{tagOptions}{List of characters. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

